+++ src/test/validation_tests.cpp
@@ -135,7 +135,7 @@ BOOST_AUTO_TEST_CASE(test_assumeutxo)
     }
 
     const auto out110 = *ExpectedAssumeutxo(110, *params);
-    BOOST_CHECK_EQUAL(out110.hash_serialized, uint256S("313082d66de67a8b79eb456728252c95acfe2f5844ae652c5da025f577fc89ab"));
+    BOOST_CHECK_EQUAL(out110.hash_serialized, uint256S("a6692e681f4819b0e21ca3b28d01fdc16045769359d14a9c4f558ca1b30736b7"));
     BOOST_CHECK_EQUAL(out110.nChainTx, (unsigned int)110);
 
     const auto out210 = *ExpectedAssumeutxo(210, *params);