+++ doc/release-notes.md
@@ -104,6 +104,10 @@ Updated RPCs
   with the `-json` option set, the following fields: `addresses`, `reqSigs` are no longer
   returned in the tx output of the response. (#20286)
 
+- The `listbanned` RPC now returns two new numeric fields: `ban_duration` and `time_remaining`.
+  Respectively, these new fields indicate the duration of a ban and the time remaining until a ban expires,
+  both in seconds. Additionally, the `ban_created` field is repositioned to come before `banned_until`. (#21602)
+
 Changes to Wallet or GUI related RPCs can be found in the GUI or Wallet section below.
 
 New RPCs
@@ -125,6 +129,9 @@ Updated settings
 Changes to Wallet or GUI related settings can be found in the GUI or Wallet section below.
 
 - Passing an invalid `-rpcauth` argument now cause bitcoind to fail to start.  (#20461)
+
+- The `getnodeaddresses` RPC now returns a "network" field indicating the
+  network type (ipv4, ipv6, onion, or i2p) for each address.  (#21594)
 
 Tools and Utilities
 -------------------